refactor: Moved doctr.documents to doctr.io
