+++ doctr/__init__.py
@@ -1,6 +1,6 @@
 from .file_utils import is_tf_available, is_torch_available
 from .version import __version__  # noqa: F401
-from . import documents
+from . import io
 from . import transforms
 from . import utils
 from . import models
