--- doctr/documents/__init__.py
@@ -1,2 +1,0 @@
-from .elements import *
-from .reader import *
