+++ doctr/io/__init__.py
@@ -1,0 +1,5 @@
+from .elements import *
+from .reader import *
+from .image import *
+from .pdf import *
+from .html import *
