+++ doctr/io/image/__init__.py
@@ -1,0 +1,8 @@
+from .base import *
+
+from doctr.file_utils import is_tf_available, is_torch_available
+
+if is_tf_available():
+    from .tensorflow import *
+elif is_torch_available():
+    from .pytorch import *  # type: ignore[misc]
