+++ doctr/io/pdf.py
@@ -7,56 +7,11 @@ import numpy as np
 import cv2
 from pathlib import Path
 import fitz
-from weasyprint import HTML
-from typing import List, Tuple, Optional, Any, Union, Sequence, Dict
+from typing import List, Tuple, Optional, Any, Dict
 
-__all__ = ['read_pdf', 'read_img', 'read_html', 'DocumentFile', 'PDF']
+from .reader import AbstractFile
 
-
-AbstractPath = Union[str, Path]
-AbstractFile = Union[AbstractPath, bytes]
-Bbox = Tuple[float, float, float, float]
-
-
-def read_img(
-    file: AbstractFile,
-    output_size: Optional[Tuple[int, int]] = None,
-    rgb_output: bool = True,
-) -> np.ndarray:
-    """Read an image file into numpy format
-
-    Example::
-        >>> from doctr.documents import read_img
-        >>> page = read_img("path/to/your/doc.jpg")
-
-    Args:
-        file: the path to the image file
-        output_size: the expected output size of each page in format H x W
-        rgb_output: whether the output ndarray channel order should be RGB instead of BGR.
-    Returns:
-        the page decoded as numpy ndarray of shape H x W x 3
-    """
-
-    if isinstance(file, (str, Path)):
-        if not Path(file).is_file():
-            raise FileNotFoundError(f"unable to access {file}")
-        img = cv2.imread(str(file), cv2.IMREAD_COLOR)
-    elif isinstance(file, bytes):
-        file = np.frombuffer(file, np.uint8)
-        img = cv2.imdecode(file, cv2.IMREAD_COLOR)
-    else:
-        raise TypeError("unsupported object type for argument 'file'")
-
-    # Validity check
-    if img is None:
-        raise ValueError("unable to read file.")
-    # Resizing
-    if isinstance(output_size, tuple):
-        img = cv2.resize(img, output_size[::-1], interpolation=cv2.INTER_LINEAR)
-    # Switch the channel order
-    if rgb_output:
-        img = cv2.cvtColor(img, cv2.COLOR_BGR2RGB)
-    return img
+__all__ = ['read_pdf', 'PDF']
 
 
 def read_pdf(file: AbstractFile, **kwargs: Any) -> fitz.Document:
@@ -129,22 +84,6 @@ def convert_page_to_numpy(
     return img
 
 
-def read_html(url: str, **kwargs: Any) -> bytes:
-    """Read a PDF file and convert it into an image in numpy format
-
-    Example::
-        >>> from doctr.documents import read_html
-        >>> doc = read_html("https://www.yoursite.com")
-
-    Args:
-        url: URL of the target web page
-    Returns:
-        decoded PDF file as a bytes stream
-    """
-
-    return HTML(url, **kwargs).write_pdf()
-
-
 class PDF:
     """PDF document template
 
@@ -204,59 +143,3 @@ class PDF:
         """
 
         return [self.get_page_artefacts(idx) for idx in range(len(self.doc))]
-
-
-class DocumentFile:
-    """Read a document from multiple extensions"""
-
-    @classmethod
-    def from_pdf(cls, file: AbstractFile, **kwargs) -> PDF:
-        """Read a PDF file
-
-        Example::
-            >>> from doctr.documents import DocumentFile
-            >>> doc = DocumentFile.from_pdf("path/to/your/doc.pdf")
-
-        Args:
-            file: the path to the PDF file or a binary stream
-        Returns:
-            a PDF document
-        """
-
-        doc = read_pdf(file, **kwargs)
-
-        return PDF(doc)
-
-    @classmethod
-    def from_url(cls, url: str, **kwargs) -> PDF:
-        """Interpret a web page as a PDF document
-
-        Example::
-            >>> from doctr.documents import DocumentFile
-            >>> doc = DocumentFile.from_url("https://www.yoursite.com")
-
-        Args:
-            url: the URL of the target web page
-        Returns:
-            a PDF document
-        """
-        pdf_stream = read_html(url)
-        return cls.from_pdf(pdf_stream, **kwargs)
-
-    @classmethod
-    def from_images(cls, files: Union[Sequence[AbstractFile], AbstractFile], **kwargs) -> List[np.ndarray]:
-        """Read an image file (or a collection of image files) and convert it into an image in numpy format
-
-        Example::
-            >>> from doctr.documents import DocumentFile
-            >>> pages = DocumentFile.from_images(["path/to/your/page1.png", "path/to/your/page2.png"])
-
-        Args:
-            files: the path to the image file or a binary stream, or a collection of those
-        Returns:
-            the list of pages decoded as numpy ndarray of shape H x W x 3
-        """
-        if isinstance(files, (str, Path, bytes)):
-            files = [files]
-
-        return [read_img(file, **kwargs) for file in files]
